+++ R/rotate_time.R
@@ -107,7 +107,7 @@ rotate_time_internal <- function(
     on.exit(DRY_RUN$deactivate())
   }
 
-  bq <- BackupQueueDateTime$new(file, format = format, backup_dir = backup_dir)
+  bq <- BackupQueueDateTime$new(file, fmt = format, backup_dir = backup_dir)
 
   # Warn if indexed backups exist
   if (BackupQueue$new(file, backup_dir = backup_dir)$has_backups){